voice_ai/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==19.10b0
boto3==1.26.142
botocore==1.29.142
braceexpand==0.1.7
cachetools==5.3.0
cdifflib==1.2.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.2
fasttext==0.9.2
ffmpy==0.3.0
filelock==3.12.0
Flask==2.2.5
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.1.2
google-auth==2.19.0
google-auth-oauthlib==1.0.0
gradio==3.32.0
gradio_client==0.2.5
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
hydra-core==1.2.0
idna==3.4
ijson==3.2.0.post0
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==6.0.4
iniconfig==2.0.0
ipadic==1.0.0
ipykernel==6.23.1
ipython==8.12.2
ipywidgets==8.0.6
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.4
kornia==0.6.12
latexcodec==2.0.1
lazy_loader==0.2
Levenshtein==0.21.0
librosa==0.10.0.post2
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.40.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab-python3==1.0.5
megatron-core==0.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-text-processing==0.1.7rc0
nemo-toolkit==1.18.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.14.0
OpenCC==0.2
orjson==3.8.14
packaging==23.1
pandas==2.0.1
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
plac==1.3.5
platformdirs==3.5.1
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
progress==1.6
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pynini==2.1.5
pyparsing==3.0.9
pypinyin==0.49.0
pypinyin-dict==0.6.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
rapidfuzz==2.13.7
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.24.0
setproctitle==1.3.2
setuptools==65.5.1
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.4.1
soxr==0.3.5
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
text-unidecode==1.3
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typed-ast==1.5.4
typer==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
wandb==0.15.3
wcwidth==0.2.6
webdataset==0.1.62
websockets==11.0.3
Werkzeug==2.3.4
wget==3.2
wheel==0.40.0
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.9.2
youtokentome==1.0.6
zipp==3.15.0
